requirements.txt
Django==4.0.0
djangorestframework==3.12.0
mysqlclient==2.2.7
python-dotenv==1.0.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.21.7
